+++ packages/contracts-communication/contracts/modules/InterchainModule.sol
@@ -1,95 +1,62 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import {InterchainModuleEvents} from "../events/InterchainModuleEvents.sol";
 import {IInterchainDB} from "../interfaces/IInterchainDB.sol";
 import {IInterchainModule} from "../interfaces/IInterchainModule.sol";
 
-import {InterchainBatch, InterchainBatchLib} from "../libs/InterchainBatch.sol";
-import {ModuleBatchLib} from "../libs/ModuleBatch.sol";
-import {VersionedPayloadLib} from "../libs/VersionedPayload.sol";
-
-import {MessageHashUtils} from "@openzeppelin/contracts/utils/cryptography/MessageHashUtils.sol";
-
 /// @notice Common logic for all Interchain Modules.
-abstract contract InterchainModule is InterchainModuleEvents, IInterchainModule {
-    using VersionedPayloadLib for bytes;
-
-    /// @notice The address of the Interchain DataBase contract: used for verifying the batches.
+abstract contract InterchainModule is IInterchainModule {
+    /// @notice The address of the Interchain DataBase contract: used for verifying the entries.
     address public immutable INTERCHAIN_DB;
 
     constructor(address interchainDB) {
         INTERCHAIN_DB = interchainDB;
     }
 
-    /// @notice Request the verification of a batch from the Interchain DataBase by the module.
-    /// If the batch is not yet finalized, the verification on destination chain will be delayed until
-    /// the finalization is done and batch root is saved on the source chain.
+    /// @notice Request the verification of an entry from the Interchain DataBase by the module.
     /// Note: a fee is paid to the module for verification, and could be retrieved by using `getModuleFee`.
-    /// Note: this will eventually trigger `InterchainDB.verifyRemoteBatch(batch)` function on destination chain,
+    /// Note: this will eventually trigger `InterchainDB.verifyRemoteEntry(entry)` function on destination chain,
     /// with no guarantee of ordering.
     /// @dev Could be only called by the Interchain DataBase contract.
     /// @param dstChainId       The chain id of the destination chain
-    /// @param batchNonce       The nonce of the batch on the source chain
-    /// @param versionedBatch   The versioned batch to verify
-    function requestBatchVerification(
-        uint64 dstChainId,
-        uint64 batchNonce,
-        bytes calldata versionedBatch
-    )
-        external
-        payable
-    {
+    /// @param versionedEntry   The versioned entry to verify
+    function requestEntryVerification(uint64 dstChainId, bytes memory versionedEntry) external payable {
         if (msg.sender != INTERCHAIN_DB) {
             revert InterchainModule__CallerNotInterchainDB(msg.sender);
         }
         if (dstChainId == block.chainid) {
             revert InterchainModule__ChainIdNotRemote(dstChainId);
         }
-        uint256 requiredFee = _getModuleFee(dstChainId, batchNonce);
+        uint256 requiredFee = _getModuleFee(dstChainId);
         if (msg.value < requiredFee) {
             revert InterchainModule__FeeAmountBelowMin({feeAmount: msg.value, minRequired: requiredFee});
         }
-        bytes memory moduleData = _fillModuleData(dstChainId, batchNonce);
-        bytes memory encodedBatch = ModuleBatchLib.encodeVersionedModuleBatch(versionedBatch, moduleData);
-        bytes32 ethSignedBatchHash = MessageHashUtils.toEthSignedMessageHash(keccak256(encodedBatch));
-        _requestVerification(dstChainId, encodedBatch);
-        emit BatchVerificationRequested(dstChainId, encodedBatch, ethSignedBatchHash);
+        // Note: we don't emit an event here, the derived contract could emit an event if needed.
+        _relayDBEntry(dstChainId, versionedEntry);
     }
 
-    /// @notice Get the Module fee for verifying a batch on the specified destination chain.
+    /// @notice Get the Module fee for verifying an entry on the specified destination chain.
     /// @param dstChainId   The chain id of the destination chain
-    /// @param dbNonce      The database nonce of the batch on the source chain
-    function getModuleFee(uint64 dstChainId, uint64 dbNonce) external view returns (uint256) {
-        return _getModuleFee(dstChainId, dbNonce);
+    function getModuleFee(uint64 dstChainId) external view returns (uint256) {
+        return _getModuleFee(dstChainId);
     }
 
-    /// @dev Should be called once the Module has verified the batch and needs to signal this
+    /// @dev Should be called once the Module has verified the entry and needs to signal this
     /// to the InterchainDB.
-    function _verifyBatch(bytes memory encodedModuleBatch) internal {
-        (bytes memory versionedBatch, bytes memory moduleData) =
-            ModuleBatchLib.decodeVersionedModuleBatch(encodedModuleBatch);
-        InterchainBatch memory batch = InterchainBatchLib.decodeBatchFromMemory(versionedBatch.getPayloadFromMemory());
-        if (batch.srcChainId == block.chainid) {
-            revert InterchainModule__ChainIdNotRemote(batch.srcChainId);
-        }
-        IInterchainDB(INTERCHAIN_DB).verifyRemoteBatch(versionedBatch);
-        _receiveModuleData(batch.srcChainId, batch.dbNonce, moduleData);
-        emit BatchVerified(
-            batch.srcChainId, encodedModuleBatch, MessageHashUtils.toEthSignedMessageHash(keccak256(encodedModuleBatch))
-        );
+    function _verifyRemoteEntry(bytes memory versionedEntry) internal {
+        IInterchainDB(INTERCHAIN_DB).verifyRemoteEntry(versionedEntry);
     }
 
     // solhint-disable no-empty-blocks
-    /// @dev Internal logic to request the verification of an batch on the destination chain.
-    function _requestVerification(uint64 dstChainId, bytes memory encodedBatch) internal virtual {}
+    /// @dev Internal logic to relay a DB entry to the destination chain.
+    /// Following checks have been done at this point:
+    /// - Entry is a valid versioned entry coming from the Interchain DataBase.
+    /// - Enough fees have been paid for the verification.
+    ///
+    /// Derived contracts should implement the logic so that eventually the destination counterpart
+    /// of this module calls `_verifyRemoteEntry(versionedEntry)`.
+    function _relayDBEntry(uint64 dstChainId, bytes memory versionedEntry) internal virtual;
 
-    /// @dev Internal logic to fill the module data for the specified destination chain.
-    function _fillModuleData(uint64 dstChainId, uint64 dbNonce) internal virtual returns (bytes memory) {}
-
-    /// @dev Internal logic to handle the auxiliary module data relayed from the remote chain.
-    function _receiveModuleData(uint64 srcChainId, uint64 dbNonce, bytes memory moduleData) internal virtual {}
-
-    /// @dev Internal logic to get the module fee for verifying an batch on the specified destination chain.
-    function _getModuleFee(uint64 dstChainId, uint64 dbNonce) internal view virtual returns (uint256);
+    /// @dev Internal logic to get the module fee for verifying an entry on the specified destination chain.
+    function _getModuleFee(uint64 dstChainId) internal view virtual returns (uint256);
 }