+++ packages/contracts-communication/contracts/modules/SynapseModule.sol
@@ -6,7 +6,10 @@ import {InterchainModule} from "./InterchainModule.sol";
 import {SynapseModuleEvents} from "../events/SynapseModuleEvents.sol";
 import {ISynapseGasOracle} from "../interfaces/ISynapseGasOracle.sol";
 import {ISynapseModule} from "../interfaces/ISynapseModule.sol";
+import {InterchainEntry, InterchainEntryLib} from "../libs/InterchainEntry.sol";
+import {ModuleEntryLib} from "../libs/ModuleEntry.sol";
 import {ThresholdECDSA} from "../libs/ThresholdECDSA.sol";
+import {VersionedPayloadLib} from "../libs/VersionedPayload.sol";
 
 import {ClaimableFees} from "../fees/ClaimableFees.sol";
 
@@ -14,13 +17,15 @@ import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {MessageHashUtils} from "@openzeppelin/contracts/utils/cryptography/MessageHashUtils.sol";
 
 contract SynapseModule is InterchainModule, ClaimableFees, Ownable, SynapseModuleEvents, ISynapseModule {
+    using VersionedPayloadLib for bytes;
+
     // TODO: make sure this is a good enough default value
     uint256 public constant DEFAULT_VERIFY_GAS_LIMIT = 100_000;
 
     /// @dev Struct to hold the verifiers and the threshold for the module.
     ThresholdECDSA internal _verifiers;
 
-    /// @dev Gas limit for the verifyBatch function on the remote chain.
+    /// @dev Gas limit for the verifyEntry function on the remote chain.
     mapping(uint64 chainId => uint256 gasLimit) internal _verifyGasLimit;
     /// @dev Hash of the last gas data sent to the remote chain.
     mapping(uint64 chainId => bytes32 gasDataHash) internal _lastGasDataHash;
@@ -109,7 +114,7 @@ contract SynapseModule is InterchainModule, ClaimableFees, Ownable, SynapseModul
         emit GasOracleSet(gasOracle_);
     }
 
-    /// @notice Sets the estimated gas limit for verifying a batch on the given chain.
+    /// @notice Sets the estimated gas limit for verifying an entry on the given chain.
     /// @dev Could be only called by the owner.
     /// @param chainId      The chain ID for which to set the gas limit
     /// @param gasLimit     The new gas limit for the verification on the specified chain
@@ -120,15 +125,22 @@ contract SynapseModule is InterchainModule, ClaimableFees, Ownable, SynapseModul
 
     // ══════════════════════════════════════════════ PERMISSIONLESS ═══════════════════════════════════════════════════
 
-    /// @notice Verifies a batch from the remote chain using a set of verifier signatures.
-    /// If the threshold is met, the batch will be marked as verified in the Interchain DataBase.
+    /// @notice Verifies an entry from the remote chain using a set of verifier signatures.
+    /// If the threshold is met, the entry will be marked as verified in the Interchain DataBase.
     /// @dev List of recovered signers from the signatures must be sorted in the ascending order.
-    /// @param encodedBatch The encoded batch to verify
-    /// @param signatures   Signatures used to verify the batch, concatenated
-    function verifyRemoteBatch(bytes calldata encodedBatch, bytes calldata signatures) external {
-        bytes32 ethSignedHash = MessageHashUtils.toEthSignedMessageHash(keccak256(encodedBatch));
+    /// @param encodedEntry The encoded entry to verify
+    /// @param signatures   Signatures used to verify the entry, concatenated
+    function verifyRemoteEntry(bytes calldata encodedEntry, bytes calldata signatures) external {
+        bytes32 ethSignedHash = MessageHashUtils.toEthSignedMessageHash(keccak256(encodedEntry));
         _verifiers.verifySignedHash(ethSignedHash, signatures);
-        _verifyBatch(encodedBatch);
+        (bytes memory versionedEntry, bytes memory data) = ModuleEntryLib.decodeVersionedModuleEntry(encodedEntry);
+        InterchainEntry memory entry = InterchainEntryLib.decodeEntryFromMemory(versionedEntry.getPayloadFromMemory());
+        if (entry.srcChainId == block.chainid) {
+            revert InterchainModule__ChainIdNotRemote(entry.srcChainId);
+        }
+        _verifyRemoteEntry(versionedEntry);
+        emit EntryVerified(entry.srcChainId, encodedEntry, ethSignedHash);
+        _receiveModuleData(entry.srcChainId, entry.dbNonce, data);
     }
 
     // ═══════════════════════════════════════════════════ VIEWS ═══════════════════════════════════════════════════════
@@ -148,7 +160,7 @@ contract SynapseModule is InterchainModule, ClaimableFees, Ownable, SynapseModul
         return _verifiers.getThreshold();
     }
 
-    /// @notice Returns the estimated gas limit for verifying a batch on the given chain.
+    /// @notice Returns the estimated gas limit for verifying an entry on the given chain.
     /// Note: this defaults to DEFAULT_VERIFY_GAS_LIMIT if not set.
     function getVerifyGasLimit(uint64 chainId) public view override returns (uint256 gasLimit) {
         gasLimit = _verifyGasLimit[chainId];
@@ -194,15 +206,22 @@ contract SynapseModule is InterchainModule, ClaimableFees, Ownable, SynapseModul
         // No op, as the claimable amount is tracked as the contract balance
     }
 
+    /// @dev Internal logic to request the verification of an entry on the destination chain.
+    /// Following checks have been done at this point:
+    /// - Entry is a valid versioned entry coming from the Interchain DataBase.
+    /// - Enough fees have been paid for the verification.
+    ///
+    /// Derived contracts should implement the logic to relay the entry to the destination chain:
+    /// the destination module counterpart should call `db.verifyRemoteEntry(versionedEntry)`.
+    function _relayDBEntry(uint64 dstChainId, bytes memory versionedEntry) internal override {
+        bytes memory moduleData = _fillModuleData(dstChainId);
+        bytes memory encodedEntry = ModuleEntryLib.encodeVersionedModuleEntry(versionedEntry, moduleData);
+        bytes32 ethSignedEntryHash = MessageHashUtils.toEthSignedMessageHash(keccak256(encodedEntry));
+        emit EntryVerificationRequested(dstChainId, encodedEntry, ethSignedEntryHash);
+    }
+
     /// @dev Internal logic to fill the module data for the specified destination chain.
-    function _fillModuleData(
-        uint64 dstChainId,
-        uint64 // dbNonce
-    )
-        internal
-        override
-        returns (bytes memory moduleData)
-    {
+    function _fillModuleData(uint64 dstChainId) internal returns (bytes memory moduleData) {
         moduleData = _getSynapseGasOracle().getLocalGasData();
         // Exit early if data is empty
         if (moduleData.length == 0) {
@@ -219,7 +238,7 @@ contract SynapseModule is InterchainModule, ClaimableFees, Ownable, SynapseModul
     }
 
     /// @dev Internal logic to handle the auxiliary module data relayed from the remote chain.
-    function _receiveModuleData(uint64 srcChainId, uint64 dbNonce, bytes memory moduleData) internal override {
+    function _receiveModuleData(uint64 srcChainId, uint64 dbNonce, bytes memory moduleData) internal {
         // Exit early if data is empty
         if (moduleData.length == 0) {
             return;
@@ -235,22 +254,14 @@ contract SynapseModule is InterchainModule, ClaimableFees, Ownable, SynapseModul
 
     // ══════════════════════════════════════════════ INTERNAL VIEWS ═══════════════════════════════════════════════════
 
-    /// @dev Internal logic to get the module fee for verifying an batch on the specified destination chain.
-    function _getModuleFee(
-        uint64 dstChainId,
-        uint64 // dbNonce
-    )
-        internal
-        view
-        override
-        returns (uint256)
-    {
-        // On the remote chain the verifyRemoteBatch(batch, signatures) function will be called.
+    /// @dev Internal logic to get the module fee for verifying an entry on the specified destination chain.
+    function _getModuleFee(uint64 dstChainId) internal view override returns (uint256) {
+        // On the remote chain the verifyRemoteEntry(entry, signatures) function will be called.
         // We need to figure out the calldata size for the remote call.
-        // selector (4 bytes) + batch + signatures
-        // batch is 32 (length) + 32*3 (fields) = 128
+        // selector (4 bytes) + entry + signatures
+        // entry is 32 (length) + 32*3 (fields) = 128
         // signatures: 32 (length) + 65*threshold (padded up to be a multiple of 32 bytes)
-        // Total formula is: 4 + 32 (batch offset) + 32 (signatures offset) + 128 + 32
+        // Total formula is: 4 + 32 (entry offset) + 32 (signatures offset) + 128 + 32
         return _getSynapseGasOracle().estimateTxCostInLocalUnits({
             remoteChainId: dstChainId,
             gasLimit: getVerifyGasLimit(dstChainId),