+++ packages/contracts-communication/test/harnesses/InterchainTransactionLibHarness.sol
@@ -11,7 +11,6 @@ contract InterchainTransactionLibHarness {
         uint64 dstChainId,
         bytes32 dstReceiver,
         uint64 dbNonce,
-        uint64 entryIndex,
         bytes memory options,
         bytes memory message
     )
@@ -20,7 +19,7 @@ contract InterchainTransactionLibHarness {
         returns (InterchainTransaction memory transaction)
     {
         return InterchainTransactionLib.constructLocalTransaction(
-            srcSender, dstChainId, dstReceiver, dbNonce, entryIndex, options, message
+            srcSender, dstChainId, dstReceiver, dbNonce, options, message
         );
     }
 
@@ -36,23 +35,14 @@ contract InterchainTransactionLibHarness {
         return InterchainTransactionLib.payloadSize(optionsLen, messageLen);
     }
 
-    function encodeTxHeader(
-        uint64 srcChainId,
-        uint64 dstChainId,
-        uint64 dbNonce,
-        uint64 entryIndex
-    )
-        external
-        pure
-        returns (ICTxHeader)
-    {
-        return InterchainTransactionLib.encodeTxHeader(srcChainId, dstChainId, dbNonce, entryIndex);
+    function encodeTxHeader(uint64 srcChainId, uint64 dstChainId, uint64 dbNonce) external pure returns (ICTxHeader) {
+        return InterchainTransactionLib.encodeTxHeader(srcChainId, dstChainId, dbNonce);
     }
 
     function decodeTxHeader(ICTxHeader header)
         external
         pure
-        returns (uint64 srcChainId, uint64 dstChainId, uint64 dbNonce, uint64 entryIndex)
+        returns (uint64 srcChainId, uint64 dstChainId, uint64 dbNonce)
     {
         return InterchainTransactionLib.decodeTxHeader(header);
     }