+++ packages/synapse-interface/components/HoverTooltip.tsx
@@ -1,21 +1,35 @@
 import React, { useState } from 'react'
 
-export const HoverTooltip = ({ children, hoverContent }) => {
+export const HoverTooltip = ({
+  children,
+  hoverContent,
+  isActive = true,
+}: {
+  children: React.ReactNode
+  hoverContent: React.ReactNode
+  isActive?: boolean
+}) => {
   const [showTooltip, setShowTooltip] = useState(false)
 
   const activateTooltip = () => setShowTooltip(true)
   const hideTooltip = () => setShowTooltip(false)
 
-  return (
-    <div
-      onMouseEnter={activateTooltip}
-      onMouseLeave={hideTooltip}
-      className="relative"
-    >
-      {children}
-      <Tooltip isHovered={showTooltip}>{hoverContent}</Tooltip>
-    </div>
-  )
+  if (!isActive) {
+    return <div>{children}</div>
+  } else {
+    return (
+      <div
+        onMouseEnter={activateTooltip}
+        onMouseLeave={hideTooltip}
+        className="relative w-fit"
+      >
+        {children}
+        {hoverContent ? (
+          <Tooltip isHovered={showTooltip}>{hoverContent}</Tooltip>
+        ) : null}
+      </div>
+    )
+  }
 }
 
 const Tooltip = ({