+++ packages/contracts-communication/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.1](https://github.com/synapsecns/sanguine/compare/contracts-communication@1.7.0...contracts-communication@1.7.1) (2024-05-12)
+
+**Note:** Version bump only for package contracts-communication
+
+
+
+
+
+# [1.7.0](https://github.com/synapsecns/sanguine/compare/contracts-communication@1.6.2...contracts-communication@1.7.0) (2024-05-11)
+
+
+### Features
+
+* **contracts-communication:** default values for modules/required respones ([#2612](https://github.com/synapsecns/sanguine/issues/2612)) ([2fcf06d](https://github.com/synapsecns/sanguine/commit/2fcf06dc2f74d610263973c028e233402e339312))
+
+
+
+
+
+## [1.6.2](https://github.com/synapsecns/sanguine/compare/contracts-communication@1.6.1...contracts-communication@1.6.2) (2024-05-10)
+
+
+### Bug Fixes
+
+* **contracts-communication:** default settings for Guard service ([#2605](https://github.com/synapsecns/sanguine/issues/2605)) ([f4ee9d0](https://github.com/synapsecns/sanguine/commit/f4ee9d0260300289c5708586b5e75a53961a8ef2))
+
+
+
+
+
+## [1.6.1](https://github.com/synapsecns/sanguine/compare/contracts-communication@1.6.0...contracts-communication@1.6.1) (2024-05-09)
+
+
+### Bug Fixes
+
+* **contracts-communication:** interchain module ignore nonces ([#2604](https://github.com/synapsecns/sanguine/issues/2604)) ([be4351b](https://github.com/synapsecns/sanguine/commit/be4351be2a5954e6d2c11471b5e93de99b3e42b6))
+
+
+
+
+
+# [1.6.0](https://github.com/synapsecns/sanguine/compare/contracts-communication@1.5.4...contracts-communication@1.6.0) (2024-05-09)
+
+
+### Features
+
+* **contracts-communication:** remove batching ([#2599](https://github.com/synapsecns/sanguine/issues/2599)) ([e6588f7](https://github.com/synapsecns/sanguine/commit/e6588f7dab0d4314e1c18b834708264177abdc7a))
+
+
+
+
+
 ## [1.5.4](https://github.com/synapsecns/sanguine/compare/contracts-communication@1.5.3...contracts-communication@1.5.4) (2024-05-08)
 
 **Note:** Version bump only for package contracts-communication