+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.5.1...@synapsecns/sdk-router@0.6.0) (2024-05-09)
+
+
+### Features
+
+* **widget:** suppress console errors ([#2594](https://github.com/synapsecns/sanguine/issues/2594)) ([925447b](https://github.com/synapsecns/sanguine/commit/925447badd2065444d5df7cc14a6e6c56e10355f)), closes [#2591](https://github.com/synapsecns/sanguine/issues/2591) [#2593](https://github.com/synapsecns/sanguine/issues/2593)
+
+
+
+
+
 ## [0.5.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.5.0...@synapsecns/sdk-router@0.5.1) (2024-05-08)